devops-ai-backend/app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
anthropic==0.7.8
pydantic==2.4.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23